# src/main/java/com/reece/addressbook/api/GetContactsResponse.java
package com.reece.addressbook.api;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;

import java.util.Set;

@Data
@Builder
@EqualsAndHashCode(callSuper = false)
@NoArgsConstructor
@AllArgsConstructor
@JsonInclude(JsonInclude.Include.NON_NULL)
public class GetContactsResponse {
    private String username;
    private Long addressId;
    private Set<GetContactResponse> contacts;
}